fix(CR-2026-06-14 ci-watch Lows): trim ci/mod.rs back under its 1435 LOC ceiling
#2195 dual-1 (r4) REJECTED: the 3 ci/mod.rs fixes pushed the grandfathered file
from its 1435 KNOWN_OVERSIZED ceiling to 1463, tripping
file_size_invariant::mcp_handler_files_under_max_loc. Per the invariant (and the
lead's retraction of the earlier ceiling-raise approval), grandfathered files may
SHRINK but must not grow past their ceiling — so the fix is to trim, NOT to raise
the recorded ceiling. file_size_invariant.rs is left UNTOUCHED.

Trims (logic of all 6 ci-watch fixes preserved; r4 confirmed them correct):
- compute_next_poll_eta moved to a sibling module src/mcp/handlers/ci/poll_eta.rs
  and re-exported (pub(crate) use) — the same LOC-relief pattern ci/mod.rs already
  uses for merge_freshness; callers and the in-module repro are unchanged. NOT the
  larger ci/mod.rs split (tracked separately).
- handle_release_repo: hoisted the prune-or-warn out of the two identical
  fallback arms to run ONCE after the match (success arm early-returns), deduping
  the warn and the prune; the tracing::warn (leak surfacing), prune, and
  remove_dir_all all remain in the body.
- Tightened the verbose CR-2026-06-14 comment blocks (eta, next_after_ci).

Result: ci/mod.rs 1463 -> 1433 (<= 1435 ceiling, no buffer padding; net -2 vs
origin/main). file_size_invariant 2/2; all 6 ci-watch repros GREEN; handlers::ci
73 + ci_watch:: 216 pass; fmt + clippy --tests --features tray -D warnings clean.


Agend-Agent: fixup-dev-3
Agend-Branch: fix/cr20260614-ci-watch-lows
Agend-Issued-At: 2026-06-15T05:46:10.982348+00:00
